Move graphviz output to the Output module
